# Update ARM NEON code