The NEON algorithms have now been updated to match the SSE2 algorithm.