@Cyan4973 2019-03-13
Thanks for this excellent `NEON` support @easyaspi314 !