@Cyan4973 2019-03-13
As a "continuation quest", it would be great if it was possible to enable `NEON` automated tests on Travis CI. So far my attempts [have not been successful](https://travis-ci.org/Cyan4973/xxHash/jobs/505548115). I'm especially surprised for `aarch64`, which works fine on my VM, but not on Travis ...