@easyaspi314 2019-03-13
OK, I see the issue. It has nothing to do with the code, it has to do with Travis's GCC version which is old af.

**GCC 4.8 disables NEON support by default because it is buggy.**

```
$ aarch64-linux-gnu-gcc-4.8 -E -xc /dev/null -dM | grep NEON
$
```

One of these should work.
```
aarch64-linux-gnu-gcc -D__ARM_NEON
clang --target=aarch64-linux-gnu -march=armv8.1-a # (probably preferred because Clang 3.8 is decent)
```