@Cyan4973 2019-03-13
For information, benchmark results on my S7+termux+clang7 platform,

without "turbo" mode,
I got 5 GB/s with this patch on `XXH3_64bits()` using `NEON` code path.
This compares favorably to 3.8 GB/s for `XXH64()`,
and 3.7 GB/s for `XXH3` using the scalar code path.

With "turbo" mode, `XXH3_64bits()` with `NEON` reaches 6.1 GB/s, 
it compares to 4.6 GB/s for `XXH64()`,
and 4.7 GB/s for `XXH3` using the scalar code path.

It also passes all consistency tests.