@Cyan4973 2019-03-13
Thanks for diagnosis @easyaspi314 ,

on Travis, I updated the `yml` script for ARM tests to use the newer `Xenial` VM.
It solves the issue for `aarch64`, which now [compiles and run correctly the `NEON` code path](https://travis-ci.org/Cyan4973/xxHash/jobs/505893894#L470).

Now, only the regular ARM 32-bit `NEON` test remains to be solved.