@Cyan4973 2019-03-13
Also, for information :
The plan now is to release a new version of xxHash, `v0.7.0`,
including this new hash,
still labelled "experimental", and protected behind a macro `XXH_STATIC_LINKING_ONLY`, because it's not stable yet.
The scope is also limited to one-short functions, though [it includes both 64 and 128 bits](https://github.com/Cyan4973/xxHash/blob/xxh3/xxhash.h#L328) variants.

This will give time to collect feedback from interested users.
There are many open questions, and I will try to list them now in code comments.

A future version will complete the scope, and move `XXH3` into stable.
Complete scope must include at least streaming hash and canonical representation.

Optional future features are custom secret keys, and secret key generation.