@Cyan4973 2019-03-13
I haven't found a good solution for 32-bit ARM and `NEON`.

I'm currently suspecting an issue regarding a mix of `hardfp` and `softfp`, maybe due to dependencies on C runtime. But that does not give me a solution.

For the time being, I will therefore disable the ARM 32-bit + `NEON` test.
`NEON` code path is validated through `aarch64`, so we are not completely blind regarding its conformance.