@easyaspi314 2019-03-13
Ok. 
Also, I did some testing with Clang's ARMv7a assembly output, and if I use the normal path without the inline assembly and remove the excess `vmov` instructions by hand, I get 8.9 GB/s. The inline assembly has an unfortunate side effect of messing with Clang's interleaving, which is why the second `vzip` didn't benefit from the asm, as the serialization wasn't good for pipelining.