@Cyan4973 2019-03-13
I'm going to prepare a blog article, to present `XXH3` when it's released.
I believe it would be appropriate to mention your contribution, at least for the `NEON` code path.
I can certainly use your pseudo `@easyaspi314`, but I could also use any other name of your preference if you wish, be it your real name or a different pseudo of your choice.